+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.13.1 / FIX ME
 
 * Empty lines are silently ignored by .rash reader
+* Added Viewpoint#members
 
 # 0.13.0 / 2013-07-29